perf(Core UI): :zap: Move zoom from graph to submodule
